Do not mix source files with generated files
I moved the generated scripts for the combined and the minified library
to the root of the project, where they can be found most easily, and they
are apart from the files edited by hand in the src/ folder.

My goal is to avoid committing by mistake generated files together with
edited files in the same commit, which, by experience, tends to cause
annoying conflicts in generated files when trying to revert a commit.