Add within.js (CC0)
within project on GitHub:
https://github.com/eric-brechemier/within

The intent is to simplify the declaration of modules by the logic to export
symbols for Node.js and for the browser.

within is a factory of semi-private spaces where properties and events
can be shared. I will replace each module pattern with:

within("coquette.maryrosecook.com", function() {
  ...
});

This will allow to share symbols defined by different modules,
now part of a common space 'coquette.maryrosecook.com', more easily.