Declare all modules within 'coquette.rosemarycook.com' space
At this point, the extra logic to export the constructor for each
type of object declared either on this.Coquette in the browser or
in exports in node is still present.

I will now update unit tests to access the constructors through within(),
removing the need to have different exports for each platform.