Reorder collider update in main update loop
Currently the renderer runs before collision response happens. If a
player collides with the ground, it's preferrable for the entity to
be moved from the collision and then drawn.

In /demos/simple, when the two boxes attach, if you look closely
the two lovers don't move in unison. This commit fixes that.